Addressing PR comments and Adding TestBase
