+++ .kokoro/presubmit/samples.cfg
@@ -31,3 +31,8 @@ env_vars: {
   key: "SECRET_MANAGER_KEYS"
   value: "java-docs-samples-service-account"
 }
+
+env_vars: {
+  key: "USER_EMAIL"
+  value: "google-cloud-java-tests@java-docs-samples-tests.iam.gserviceaccount.com"
+}
